Overhauls the mechanics of schema validation
Yippie, that makes the code more lightweight and much easier to
understand.
