+++ cerberus/tests/test_assorted.py
@@ -2,9 +2,9 @@ from decimal import Decimal
 
 from pytest import mark
 
-from cerberus import TypeDefinition, Validator
+from cerberus import validator_factory, TypeDefinition, Validator
+from cerberus.base import UnconcernedValidator
 from cerberus.tests import assert_fail, assert_success
-from cerberus.utils import validator_factory
 
 
 def test_clear_cache(validator):
@@ -52,7 +52,11 @@ def test_dynamic_types():
 
 
 def test_mro():
-    assert Validator.__mro__ == (Validator, object), Validator.__mro__
+    assert Validator.__mro__ == (
+        Validator,
+        UnconcernedValidator,
+        object,
+    ), Validator.__mro__
 
 
 def test_mixin_init():
